Split task-related database functions into its own file
